+++ src/TeamAdmin.Lib/Repositories/TeamRepository.cs
@@ -133,7 +133,7 @@ namespace TeamAdmin.Lib.Repositories
             }
         }
 
-        public bool DeleteMedia(int mediaId)
+        public bool DeleteMedia(long mediaId)
         {
             using (var context = ContextFactory.Create<ClubContext>())
             {
@@ -146,7 +146,7 @@ namespace TeamAdmin.Lib.Repositories
             }
         }
 
-        public void UpdateMediaCaption(int mediaId, string newCaption)
+        public void UpdateMediaCaption(long mediaId, string newCaption)
         {
             using (var context = ContextFactory.Create<ClubContext>())
             {
@@ -159,7 +159,7 @@ namespace TeamAdmin.Lib.Repositories
             }
         }
 
-        public bool SetMediaPosition(int mediaId, int newPosition)
+        public bool SetMediaPosition(long mediaId, int newPosition)
         {
             using (var context = ContextFactory.Create<ClubContext>())
             {
